+++ app/src/main/java/com/inky/fitnesscalendar/db/Database.kt
@@ -30,7 +30,7 @@ import com.inky.fitnesscalendar.util.DATABASE_NAME
 private const val TAG = "Database"
 
 @Database(
-    version = 39,
+    version = 40,
     entities = [
         Activity::class,
         Recording::class,
@@ -69,6 +69,7 @@ private const val TAG = "Database"
         AutoMigration(from = 35, to = 36, spec = Migration35To36Spec::class),
         AutoMigration(from = 37, to = 38, spec = Migration37To38Spec::class),
         AutoMigration(from = 38, to = 39),
+        AutoMigration(from = 39, to = 40),
     ]
 )
 @TypeConverters(Converters::class)